% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_score.R
\name{calculate_score}
\alias{calculate_score}
\title{Calculate indicator scores following DERI methodology}
\usage{
calculate_score(var, inverse = FALSE)
}
\arguments{
\item{var}{Vector containing the indicator to be scored (e.g., gp_dist_km)}

\item{inverse}{Default is FALSE. Set as TRUE if higher value of the indicator
indicates lower risk}
}
\description{
As explained in the \href{https://github.com/GreaterManchesterODA/Digital-Exclusion-Risk-Index/blob/main/Version\%201.0/DERI\%20Score\%20Methodology_v1.0.md}{DERI Score Methodology}:
}
\details{
"For the DERI, the score for each indicator in each LSOA is given as a number
between 0 and 10, where 10 indicates a higher level of digital exclusion
risk. The score for each indicator aimed to show how far between the minimum
and maximum indicator values the individual indicator value was for each LSOA
located. This gives a value between 0 (the LSOA has the smallest indicator
value, therefore indicating low risk) to 10 (the LSOA has the highest
indicator value, therefore indicating high risk)."

The calculation used for each indicator (i) and each LSOA (j) in each
district (k) is:
\itemize{
\item For indicators for which higher value indicates higher risk:
10 x ((Value of indicator i for LSOA j in district k) -
(Minimum value of indicator i across all LSOAs in district k)) /
((Maximum value of i across all LSOAs in district k) -
(Minimum value of i across all LSOAs in district k))
\item For indicators for which higher value indicates lower risk:
10 x ((Maximum value of indicator i across all LSOAs in district k) -
(Value of indicator i for LSOA j in district k)) /
((Maximum value of i across all LSOAs in district k) -
(Minimum value of i across all LSOAs in district k))
}

As a result, the direction of the scale of data inputted by the function
does not necessarily match that of the data outputted, for which higher
values will always mean high risk.
}
